Correction of errors found during further analysus

+++ database/migrations/2022_02_25_193119_add_balance_to_users.php
@@ -15,7 +15,7 @@ class AddBalanceToUsers extends Migration
     {
         Schema::table('users', function (Blueprint $table) {
             //
-            $table->float('balance')->default(0.0);
+            $table->double('balance')->default(0.0);
         });
     }
